backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
asyncpg==0.30.0
PyJWT[crypto]==2.10.1
pydantic==2.11.7
pydantic[email]==2.11.7
cohere==5.18.0
python-dotenv==1.1.1
pytest==8.4.2
httpx==0.28.1
